+++ src/icons.js
@@ -1,3 +1,0 @@
-import { library } from "@fortawesome/fontawesome-svg-core";
-import { faPlus, faCircleXmark } from "@fortawesome/free-solid-svg-icons";
-library.add(faPlus, faCircleXmark);