+++ src/resources/icons/icons.js
@@ -2,8 +2,23 @@ import { library } from "@fortawesome/fontawesome-svg-core";
 import {
   faBars,
   faTimes,
+  faTimesCircle,
+  faCheckCircle,
   faCaretSquareDown,
   faCaretSquareUp,
+  faMinus,
+  faPlus,
+  faCircleXmark,
 } from "@fortawesome/free-solid-svg-icons";
 
-library.add(faBars, faTimes, faCaretSquareDown, faCaretSquareUp);
+library.add(
+  faBars,
+  faMinus,
+  faTimes,
+  faTimesCircle,
+  faCheckCircle,
+  faCaretSquareDown,
+  faCaretSquareUp,
+  faPlus,
+  faCircleXmark
+);